requirements_advanced.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
filterpy>=1.4.5
pyyaml>=6.0

# Note: Requires Python 3.8-3.12 (MediaPipe doesn't support 3.13 yet)
# Install with: pip install -r requirements_advanced.txt
